Refactor in networking package
